// src/app/modules/booking/booking.interface.ts
import { Types } from "mongoose";

export type TVehicleType =
    | "car"
    | 'truck'
    | 'SUV'
    | 'van'
    | 'motorcycle'
    | 'bus'
    | 'electricVehicle'
    | 'hybridVehicle'
    | 'bicycle'
    | 'tractor';


export type TBooking = {
    customer: Types.ObjectId;
    service: Types.ObjectId;
    slot: Types.ObjectId;
    vehicleType: TVehicleType
    vehicleBrand: string;
    vehicleModel: string;
    manufacturingYear: number;
    registrationPlate: string;
}